perf: Remove old swap components
