+++ apps/web/src/views/AddLiquidity/AddStableLiquidity/hooks/useDerivedLPInfo.ts
@@ -3,7 +3,7 @@ import { useContext, useMemo } from 'react'
 import { StableSwap } from '@pancakeswap/smart-router/evm'
 import tryParseAmount from '@pancakeswap/utils/tryParseAmount'
 
-import { StableConfigContext } from 'views/Swap/StableSwap/hooks/useStableConfig'
+import { StableConfigContext } from 'views/Swap/hooks/useStableConfig'
 import { useStableSwapInfo } from 'hooks/useStableSwapInfo'
 import { BIG_INT_ZERO } from 'config/constants/exchange'
