+++ apps/web/src/views/AddLiquidityV3/components/StablePairCard.tsx
@@ -1,7 +1,7 @@
 import { Tag } from '@pancakeswap/uikit'
 import { useTokenBalance } from 'state/wallet/hooks'
 import { useGetRemovedTokenAmounts } from 'views/RemoveLiquidity/RemoveStableLiquidity/hooks/useStableDerivedBurnInfo'
-import { LPStablePair } from 'views/Swap/StableSwap/hooks/useStableConfig'
+import { LPStablePair } from 'views/Swap/hooks/useStableConfig'
 import { LiquidityCardRow } from 'components/LiquidityCardRow'
 
 export function StablePairCard({ pair, account }: { pair: LPStablePair; account: string }) {
