--- apps/web/src/views/Swap/SmartSwap/hooks/index.ts
@@ -1,3 +1,0 @@
-export * from './useDerivedSwapInfoWithStableSwap'
-export * from './useTradeInfo'
-export * from './useIsSmartRouterBetter'
