--- apps/web/src/views/Swap/StableSwap/components/StableSwapModalFooter.tsx
@@ -1,104 +1,0 @@
-import { useTranslation } from '@pancakeswap/localization'
-import { Currency, CurrencyAmount, TradeType } from '@pancakeswap/sdk'
-import { AutoRenewIcon, Button, QuestionHelper, Text, AutoColumn } from '@pancakeswap/uikit'
-import { AutoRow, RowBetween, RowFixed } from 'components/Layout/Row'
-import { useState } from 'react'
-import { Field } from 'state/swap/actions'
-import styled from 'styled-components'
-import { formatExecutionPrice } from 'utils/exchange'
-import { StableTrade } from 'views/Swap/StableSwap/hooks/useStableTradeExactIn'
-import { StyledBalanceMaxMini, SwapCallbackError } from '../../components/styleds'
-
-const SwapModalFooterContainer = styled(AutoColumn)`
-  margin-top: 24px;
-  padding: 16px;
-  border-radius: ${({ theme }) => theme.radii.default};
-  border: 1px solid ${({ theme }) => theme.colors.cardBorder};
-  background-color: ${({ theme }) => theme.colors.background};
-`
-
-export default function StableSwapModalFooter({
-  trade,
-  slippageAdjustedAmounts,
-  isEnoughInputBalance,
-  onConfirm,
-  swapErrorMessage,
-  disabledConfirm,
-}: {
-  trade: StableTrade
-  slippageAdjustedAmounts: { [field in Field]?: CurrencyAmount<Currency> }
-  isEnoughInputBalance: boolean
-  onConfirm: () => void
-  swapErrorMessage?: string | undefined
-  disabledConfirm: boolean
-}) {
-  const { t } = useTranslation()
-  const [showInverted, setShowInverted] = useState<boolean>(false)
-
-  return (
-    <>
-      <SwapModalFooterContainer>
-        <RowBetween align="center">
-          <Text fontSize="14px">{t('Price')}</Text>
-          <Text
-            fontSize="14px"
-            style={{
-              justifyContent: 'center',
-              alignItems: 'center',
-              display: 'flex',
-              textAlign: 'right',
-              paddingLeft: '10px',
-            }}
-          >
-            {formatExecutionPrice(trade, showInverted)}
-            <StyledBalanceMaxMini onClick={() => setShowInverted(!showInverted)}>
-              <AutoRenewIcon width="14px" />
-            </StyledBalanceMaxMini>
-          </Text>
-        </RowBetween>
-
-        <RowBetween>
-          <RowFixed>
-            <Text fontSize="14px">
-              {trade.tradeType === TradeType.EXACT_INPUT ? t('Minimum received') : t('Maximum sold')}
-            </Text>
-            <QuestionHelper
-              text={t(
-                'Your transaction will revert if there is a large, unfavorable price movement before it is confirmed.',
-              )}
-              ml="4px"
-              placement="top"
-            />
-          </RowFixed>
-          <RowFixed>
-            <Text fontSize="14px">
-              {trade.tradeType === TradeType.EXACT_INPUT
-                ? slippageAdjustedAmounts[Field.OUTPUT]?.toSignificant(4) ?? '-'
-                : slippageAdjustedAmounts[Field.INPUT]?.toSignificant(4) ?? '-'}
-            </Text>
-            <Text fontSize="14px" marginLeft="4px">
-              {trade.tradeType === TradeType.EXACT_INPUT
-                ? trade.outputAmount.currency.symbol
-                : trade.inputAmount.currency.symbol}
-            </Text>
-          </RowFixed>
-        </RowBetween>
-      </SwapModalFooterContainer>
-
-      <AutoRow>
-        <Button
-          variant="primary"
-          onClick={onConfirm}
-          disabled={disabledConfirm}
-          mt="12px"
-          id="confirm-swap-or-send"
-          width="100%"
-        >
-          {trade.tradeType === TradeType.EXACT_OUTPUT && !isEnoughInputBalance ? t('Swap Anyway') : t('Confirm Swap')}
-        </Button>
-
-        {swapErrorMessage ? <SwapCallbackError error={swapErrorMessage} /> : null}
-      </AutoRow>
-    </>
-  )
-}
