--- apps/web/src/views/Swap/StableSwap/hooks/useDerivedStableSwapInfo.ts
@@ -1,95 +1,0 @@
-import { useAccount } from 'wagmi'
-import { Currency, CurrencyAmount } from '@pancakeswap/sdk'
-import { useTranslation } from '@pancakeswap/localization'
-import { isAddress } from 'utils'
-
-import tryParseAmount from '@pancakeswap/utils/tryParseAmount'
-import { Field } from 'state/swap/actions'
-import { computeSlippageAdjustedAmounts } from 'utils/exchange'
-import { useUserSlippage } from '@pancakeswap/utils/user'
-import { useCurrencyBalances } from 'state/wallet/hooks'
-import { useMemo } from 'react'
-import useStableTradeExactIn, { StableTrade } from './useStableTradeExactIn'
-import useStableTradeExactOut from './useStableTradeExactOut'
-
-// from the current swap inputs, compute the best trade and return it.
-export function useDerivedStableSwapInfo(
-  independentField: Field,
-  typedValue: string,
-  inputCurrency: Currency | undefined,
-  outputCurrency: Currency | undefined,
-): {
-  currencies: { [field in Field]?: Currency }
-  currencyBalances: { [field in Field]?: CurrencyAmount<Currency> }
-  parsedAmount: CurrencyAmount<Currency> | undefined
-  v2Trade: StableTrade | undefined
-  inputError?: string
-} {
-  const { address: account } = useAccount()
-  const { t } = useTranslation()
-
-  const to: string | null = account ?? null
-
-  const relevantTokenBalances = useCurrencyBalances(
-    account ?? undefined,
-    useMemo(() => [inputCurrency ?? undefined, outputCurrency ?? undefined], [inputCurrency, outputCurrency]),
-  )
-
-  const isExactIn: boolean = independentField === Field.INPUT
-  const parsedAmount = tryParseAmount(typedValue, (isExactIn ? inputCurrency : outputCurrency) ?? undefined)
-
-  const bestTradeExactIn = useStableTradeExactIn(isExactIn ? parsedAmount : undefined, outputCurrency ?? undefined)
-  const bestTradeExactOut = useStableTradeExactOut(inputCurrency ?? undefined, !isExactIn ? parsedAmount : undefined)
-
-  const v2Trade = isExactIn ? bestTradeExactIn : bestTradeExactOut
-
-  const currencyBalances = {
-    [Field.INPUT]: relevantTokenBalances[0],
-    [Field.OUTPUT]: relevantTokenBalances[1],
-  }
-
-  const currencies: { [field in Field]?: Currency } = {
-    [Field.INPUT]: inputCurrency ?? undefined,
-    [Field.OUTPUT]: outputCurrency ?? undefined,
-  }
-
-  let inputError: string | undefined
-  if (!account) {
-    inputError = t('Connect Wallet')
-  }
-
-  if (!parsedAmount) {
-    inputError = inputError ?? t('Enter an amount')
-  }
-
-  if (!currencies[Field.INPUT] || !currencies[Field.OUTPUT]) {
-    inputError = inputError ?? t('Select a token')
-  }
-
-  const formattedTo = isAddress(to)
-  if (!to || !formattedTo) {
-    inputError = inputError ?? t('Enter a recipient')
-  }
-
-  const [allowedSlippage] = useUserSlippage()
-
-  const slippageAdjustedAmounts = v2Trade && allowedSlippage && computeSlippageAdjustedAmounts(v2Trade, allowedSlippage)
-
-  // compare input balance to max input based on version
-  const [balanceIn, amountIn] = [
-    currencyBalances[Field.INPUT],
-    slippageAdjustedAmounts ? slippageAdjustedAmounts[Field.INPUT] : null,
-  ]
-
-  if (balanceIn && amountIn && balanceIn.lessThan(amountIn)) {
-    inputError = t('Insufficient %symbol% balance', { symbol: amountIn.currency.symbol })
-  }
-
-  return {
-    currencies,
-    currencyBalances,
-    parsedAmount,
-    v2Trade: v2Trade ?? undefined,
-    inputError,
-  }
-}
