+++ apps/web/src/views/Swap/hooks/useStableConfig.ts
